Moves get initial info, removes account route that is old
