+++ routes/login/login.js
@@ -1,62 +1,9 @@
 const db = require("../../data/config");
-const axios = require("axios");
 
 module.exports = {
-  getAPIToken,
-  getUserInfo,
   findUser,
-  findSubId,
   registerUser
 };
-
-function getAPIToken() {
-  return new Promise(async (resolve, reject) => {
-    try {
-      const body = {
-        client_id: process.env.OAUTH_CLIENT_ID,
-        client_secret: process.env.OAUTH_CLIENT_SECRET,
-        audience: process.env.OAUTH_MGR_API,
-        grant_type: "client_credentials"
-      };
-      const { data } = await axios.post(process.env.OAUTH_TOKEN_API, body);
-      resolve(data.access_token);
-    } catch (error) {
-      console.log(error);
-      reject(error);
-    }
-  });
-}
-
-function getUserInfo(token, id) {
-  return new Promise(async (resolve, reject) => {
-    try {
-      const api = process.env.OAUTH_MGR_API;
-      const url = `${api}users/${id}`;
-      const headers = {
-        authorization: `Bearer ${token}`
-      };
-
-      const { data } = await axios.get(url, { headers });
-      const info = {
-        name: data.name,
-        github: data.html_url,
-        location: data.location,
-        about: data.bio
-      };
-      resolve(info);
-    } catch (error) {
-      console.log(error);
-      reject(error);
-    }
-  });
-}
-
-function findSubId(account_id) {
-  return db("accounts")
-    .select("sub_id")
-    .where({ id: account_id })
-    .first();
-}
 
 function findUser(sub_id) {
   return db("accounts")
